+++ src/minified-generated-full-src.js
@@ -938,7 +938,7 @@ define('minified', function() {
 	}
 	
 	function delay(f, delayMs) {
-		_window.setTimeout(f, delayMs||0);
+		setTimeout(f, delayMs||0);
 	}
 	function extractNumber(v) {
 		return parseFloat(replace(v, /^[^\d-]+/));
@@ -1327,7 +1327,7 @@ define('minified', function() {
 	 * </li></ul>
 	 * 
 	 * @example A simple promise that is fulfilled after 1 second:
-	 * <pre>var p = promise();
+	 * <pre>var p = _.promise();
 	 * setTimeout(function() { 
 	 *     p(true, []); 
 	 * }, 1000);
@@ -1369,8 +1369,9 @@ define('minified', function() {
 		each(assimilatedPromises, function assimilate(promise, index) {
 			try {
 				promise['then'](function resolvePromise(v) {
-					if ((isObject(v) || isFunction(v)) && isFunction(v['then'])) {
-						assimilate(v['then'], index);
+					var then;
+					if ((isObject(v) || isFunction(v)) && isFunction(then = v['then'])) {
+						assimilate(then, index);
 					}
 					else {
 						values[index] = map(arguments, nonOp);
@@ -1472,17 +1473,18 @@ define('minified', function() {
 	    			if (isFunction(f)) {
 		   				function resolve(x) {
 		   					try {
-			   					var then;
+			   					var then, cbCalled = 0;
 				   				if ((isObject(x) || isFunction(x)) && isFunction(then = x['then'])) {
 										if (x === promise2)
 											throw new TypeError();
-										then['call'](x, resolve, function(value){promise2(false,[value]);});
+										then['call'](x, function(x) { if (!cbCalled++) resolve(x); } , function(value){ if (!cbCalled++) promise2(false,[value]);});
 				   				}
 				   				else
 				   					promise2(true, [x]);
 		   					}
 		   					catch(e) {
-		   						promise2(false, [e]);
+		   						if (!cbCalled++) 
+		   							promise2(false, [e]);
 		   					}
 		   				}
 		   				resolve(call(f, undef, values));